--- a/Jint.Tests.Test262/Language/GlobalCodeTests.cs
+++ b/Jint.Tests.Test262/Language/GlobalCodeTests.cs
@@ -1,0 +1,15 @@
+using Xunit;
+
+namespace Jint.Tests.Test262.Language
+{
+    public class GlobalCodeTests : Test262Test
+    {
+        [Theory(DisplayName = "language\\global-code")]
+        [MemberData(nameof(SourceFiles), "language\\global-code", false)]
+        [MemberData(nameof(SourceFiles), "language\\global-code", true, Skip = "Skipped")]
+        protected void GlobalCode(SourceFile sourceFile)
+        {
+            RunTestInternal(sourceFile);
+        }
+    }
+}