--- a/Jint.Tests.Test262/Language/Statements/ConstTests.cs
+++ b/Jint.Tests.Test262/Language/Statements/ConstTests.cs
@@ -4,7 +4,7 @@ namespace Jint.Tests.Test262.Language.Statements
 {
     public class ConstTests : Test262Test
     {
-        [Theory(DisplayName = "language\\statements\\const", Skip = "Skipped")]
+        [Theory(DisplayName = "language\\statements\\const")]
         [MemberData(nameof(SourceFiles), "language\\statements\\const", false)]
         [MemberData(nameof(SourceFiles), "language\\statements\\const", true, Skip = "Skipped")]
         protected void For(SourceFile sourceFile)