* move hoisting to Jint's side
* add new binding initialization

fixes #379 
fixes #583
fixes #714 
fixes #732