@lahma 2020-05-23
Current perf, need to investigate the biggest losses, but as predicted this will be a bit slower than before due to new complexity.

## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|467.2 ms|0.67 ms|12000.0000|1000.0000|0.0000|50.03 MB|
| **New** |	|	| **545.3 ms (+17%)** | **1.52 ms** | **13000.0000 (+8%)** | **1000.0000 (0%)** | **0.0000** | **54.34 MB (+9%)** |
| Old |Run|3d-morph|381.7 ms|3.47 ms|12000.0000|2000.0000|0.0000|77.75 MB|
| **New** |	|	| **469.1 ms (+23%)** | **3.35 ms** | **17000.0000 (+42%)** | **3000.0000 (+50%)** | **0.0000** | **108.33 MB (+39%)** |
| Old |Run|3d-raytrace|348.3 ms|3.84 ms|21000.0000|3000.0000|0.0000|85.9 MB|
| **New** |	|	| **396.7 ms (+14%)** | **3.25 ms** | **23000.0000 (+10%)** | **3000.0000 (0%)** | **0.0000** | **96.04 MB (+12%)** |
| Old |Run|access-binary-trees|160.3 ms|0.96 ms|14500.0000|500.0000|0.0000|60.6 MB|
| **New** |	|	| **192.8 ms (+20%)** | **2.10 ms** | **21000.0000 (+45%)** | **1000.0000 (+100%)** | **0.0000** | **86.42 MB (+43%)** |
| Old |Run|access-fannkuch|1,191.8 ms|14.19 ms|9000.0000|0.0000|0.0000|37.94 MB|
| **New** |	|	| **1,323.5 ms (+11%)** | **12.61 ms** | **9000.0000 (0%)** | **0.0000** | **0.0000** | **37.95 MB (0%)** |
| Old |Run|access-nbody|453.2 ms|5.46 ms|15000.0000|0.0000|0.0000|62.03 MB|
| **New** |	|	| **460.8 ms (+2%)** | **2.58 ms** | **16000.0000 (+7%)** | **0.0000** | **0.0000** | **65.51 MB (+6%)** |
| Old |Run|access-nsieve|491.1 ms|4.51 ms|17000.0000|5000.0000|1000.0000|85.24 MB|
| **New** |	|	| **519.2 ms (+6%)** | **2.65 ms** | **17000.0000 (0%)** | **5000.0000 (0%)** | **1000.0000 (0%)** | **85.24 MB (0%)** |
| Old |Run|bitop(...)-byte [24]|281.4 ms|2.81 ms|13000.0000|0.0000|0.0000|53.69 MB|
| **New** |	|	| **325.0 ms (+15%)** | **3.81 ms** | **18000.0000 (+38%)** | **0.0000** | **0.0000** | **73.22 MB (+36%)** |
| Old |Run|bitops-bits-in-byte|482.8 ms|3.98 ms|8000.0000|0.0000|0.0000|34.83 MB|
| **New** |	|	| **564.3 ms (+17%)** | **2.01 ms** | **12000.0000 (+50%)** | **0.0000** | **0.0000** | **48.51 MB (+39%)** |
| Old |Run|bitops-bitwise-and|289.3 ms|1.10 ms|22000.0000|0.0000|0.0000|91.26 MB|
| **New** |	|	| **585.7 ms (+102%)** | **1.12 ms** | **36000.0000 (+64%)** | **0.0000** | **0.0000** | **146.19 MB (+60%)** |
| Old |Run|bitops-nsieve-bits|458.9 ms|2.16 ms|24000.0000|1000.0000|0.0000|98.89 MB|
| **New** |	|	| **500.0 ms (+9%)** | **4.59 ms** | **24000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **99.81 MB (+1%)** |
| Old |Run|contr(...)rsive [21]|216.3 ms|2.11 ms|19000.0000|2000.0000|0.0000|79.6 MB|
| **New** |	|	| **284.0 ms (+31%)** | **0.71 ms** | **34000.0000 (+79%)** | **1000.0000 (-50%)** | **0.0000** | **139.54 MB (+75%)** |
| Old |Run|crypto-aes|342.5 ms|2.84 ms|6000.0000|1000.0000|0.0000|26.16 MB|
| **New** |	|	| **371.8 ms (+9%)** | **1.08 ms** | **7000.0000 (+17%)** | **1000.0000 (0%)** | **0.0000** | **32.73 MB (+25%)** |
| Old |Run|crypto-md5|196.9 ms|1.56 ms|18000.0000|2000.0000|0.0000|76.92 MB|
| **New** |	|	| **234.6 ms (+19%)** | **1.34 ms** | **26000.0000 (+44%)** | **1000.0000 (-50%)** | **0.0000** | **107.2 MB (+39%)** |
| Old |Run|crypto-sha1|199.3 ms|1.62 ms|15000.0000|1000.0000|0.0000|64.66 MB|
| **New** |	|	| **230.5 ms (+16%)** | **0.73 ms** | **23000.0000 (+53%)** | **1000.0000 (0%)** | **0.0000** | **94.47 MB (+46%)** |
| Old |Run|date-format-tofte|242.2 ms|0.95 ms|16000.0000|0.0000|0.0000|65.77 MB|
| **New** |	|	| **284.6 ms (+18%)** | **2.08 ms** | **17000.0000 (+6%)** | **1000.0000** | **0.0000** | **70.68 MB (+7%)** |
| Old |Run|date-format-xparb|117.5 ms|0.24 ms|5600.0000|200.0000|0.0000|23.92 MB|
| **New** |	|	| **137.4 ms (+17%)** | **0.72 ms** | **8000.0000 (+43%)** | **1000.0000 (+400%)** | **0.0000** | **34.68 MB (+45%)** |
| Old |Run|math-cordic|686.0 ms|3.02 ms|25000.0000|0.0000|0.0000|100.14 MB|
| **New** |	|	| **799.7 ms (+17%)** | **3.64 ms** | **38000.0000 (+52%)** | **0.0000** | **0.0000** | **154.31 MB (+54%)** |
| Old |Run|math-partial-sums|255.7 ms|2.35 ms|12000.0000|0.0000|0.0000|48.97 MB|
| **New** |	|	| **351.6 ms (+38%)** | **1.87 ms** | **24000.0000 (+100%)** | **0.0000** | **0.0000** | **97.42 MB (+99%)** |
| Old |Run|math-spectral-norm|217.5 ms|0.95 ms|12000.0000|0.0000|0.0000|51.68 MB|
| **New** |	|	| **261.5 ms (+20%)** | **1.02 ms** | **17000.0000 (+42%)** | **0.0000** | **0.0000** | **70.43 MB (+36%)** |
| Old |Run|regexp-dna|406.6 ms|1.14 ms|0.0000|0.0000|0.0000|18.3 MB|
| **New** |	|	| **417.6 ms (+3%)** | **1.42 ms** | **0.0000** | **0.0000** | **0.0000** | **18.31 MB (0%)** |
| Old |Run|string-base64|189.3 ms|1.19 ms|3000.0000|1000.0000|0.0000|13.13 MB|
| **New** |	|	| **207.2 ms (+9%)** | **1.05 ms** | **5000.0000 (+67%)** | **1000.0000 (0%)** | **0.0000** | **22.14 MB (+69%)** |
| Old |Run|string-fasta|378.9 ms|2.16 ms|29000.0000|0.0000|0.0000|117.41 MB|
| **New** |	|	| **497.3 ms (+31%)** | **2.34 ms** | **45000.0000 (+55%)** | **0.0000** | **0.0000** | **180.5 MB (+54%)** |
| Old |Run|string-tagcloud|165.7 ms|1.02 ms|8000.0000|1000.0000|0.0000|46.19 MB|
| **New** |	|	| **175.4 ms (+6%)** | **1.26 ms** | **11000.0000 (+38%)** | **1000.0000 (0%)** | **0.0000** | **56.36 MB (+22%)** |
| Old |Run|string-unpack-code|167.0 ms|1.51 ms|15000.0000|4000.0000|0.0000|71.94 MB|
| **New** |	|	| **180.7 ms (+8%)** | **0.40 ms** | **20000.0000 (+33%)** | **6000.0000 (+50%)** | **1000.0000** | **88.63 MB (+23%)** |
| Old |Run|strin(...)input [21]|1,089.1 ms|6.95 ms|1099000.0000|1080000.0000|1079000.0000|6207.8 MB|
| **New** |	|	| **1,126.2 ms (+3%)** | **10.52 ms** | **1119000.0000 (+2%)** | **1096000.0000 (+1%)** | **1095000.0000 (+1%)** | **6221.02 MB (0%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Slice|100|411.4 μs|1.82 μs|180.6641|0.0000|0.0000|738.28 KB|
| **New** |	|	| **461.5 μs (+12%)** | **1.59 μs** | **181.6406 (+1%)** | **0.0000** | **0.0000** | **742.97 KB (+1%)** |
| Old |Concat|100|595.7 μs|7.27 μs|216.7969|0.0000|0.0000|885.94 KB|
| **New** |	|	| **662.8 μs (+11%)** | **4.80 μs** | **217.7734 (0%)** | **0.0000** | **0.0000** | **890.63 KB (+1%)** |
| Old |Unshift|100|17,431.4 μs|174.88 μs|3843.7500|0.0000|0.0000|15733.59 KB|
| **New** |	|	| **19,987.5 μs (+15%)** | **55.42 μs** | **4031.2500 (+5%)** | **0.0000** | **0.0000** | **16519.53 KB (+5%)** |
| Old |Push|100|8,622.7 μs|35.72 μs|625.0000|0.0000|0.0000|2585.16 KB|
| **New** |	|	| **12,979.4 μs (+51%)** | **94.59 μs** | **812.5000 (+30%)** | **0.0000** | **0.0000** | **3371.09 KB (+30%)** |
| Old |Index|100|9,730.1 μs|49.02 μs|812.5000|0.0000|0.0000|3321.88 KB|
| **New** |	|	| **14,163.2 μs (+46%)** | **27.95 μs** | **1000.0000 (+23%)** | **0.0000** | **0.0000** | **4119.53 KB (+24%)** |
| Old |Map|100|3,745.5 μs|27.53 μs|796.8750|0.0000|0.0000|3255.47 KB|
| **New** |	|	| **4,425.1 μs (+18%)** | **33.07 μs** | **1125.0000 (+41%)** | **0.0000** | **0.0000** | **4610.96 KB (+42%)** |
| Old |Apply|100|624.8 μs|3.41 μs|212.8906|0.0000|0.0000|872.66 KB|
| **New** |	|	| **688.4 μs (+10%)** | **2.78 μs** | **215.8203 (+1%)** | **0.0000** | **0.0000** | **883.59 KB (+1%)** |
| Old |JsonStringifyParse|100|4,526.5 μs|32.79 μs|1296.8750|0.0000|0.0000|5320.32 KB|
| **New** |	|	| **4,474.9 μs (-1%)** | **16.93 μs** | **1304.6875 (+1%)** | **0.0000** | **0.0000** | **5337.51 KB (0%)** |
| Old |FilterWithString|100|40,702.7 μs|443.64 μs|6461.5385|0.0000|0.0000|26635.16 KB|
| **New** |	|	| **49,034.3 μs (+20%)** | **428.07 μs** | **10272.7273 (+59%)** | **0.0000** | **0.0000** | **42323.44 KB (+59%)** |


## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|59.16 ms|0.151 ms|1444.4444|444.4444|0.0000|8134.62 KB|
| **New** |	|	| **64.07 ms (+8%)** | **0.090 ms** | **1625.0000 (+13%)** | **250.0000 (-44%)** | **0.0000** | **8996.02 KB (+11%)** |
| Old |Run|dromaeo-core-eval|15.05 ms|0.057 ms|125.0000|0.0000|0.0000|515.16 KB|
| **New** |	|	| **17.79 ms (+18%)** | **0.208 ms** | **125.0000 (0%)** | **0.0000** | **0.0000** | **525.7 KB (+2%)** |
| Old |Run|dromaeo-object-array|157.99 ms|0.683 ms|42000.0000|2000.0000|0.0000|177537.93 KB|
| **New** |	|	| **188.86 ms (+20%)** | **1.921 ms** | **44000.0000 (+5%)** | **2000.0000 (0%)** | **0.0000** | **184206.28 KB (+4%)** |
| Old |Run|droma(...)egexp [21]|657.84 ms|13.011 ms|46000.0000|26000.0000|17000.0000|373592.02 KB|
| **New** |	|	| **711.54 ms (+8%)** | **14.217 ms** | **48000.0000 (+4%)** | **26000.0000 (0%)** | **15000.0000 (-12%)** | **392946.5 KB (+5%)** |
| Old |Run|droma(...)tring [21]|520.41 ms|6.181 ms|104000.0000|67000.0000|65000.0000|1355546.3 KB|
| **New** |	|	| **631.37 ms (+21%)** | **5.853 ms** | **123000.0000 (+18%)** | **76000.0000 (+13%)** | **73000.0000 (+12%)** | **1404181.62 KB (+4%)** |
| Old |Run|droma(...)ase64 [21]|120.61 ms|0.853 ms|2000.0000|400.0000|0.0000|8920.38 KB|
| **New** |	|	| **135.09 ms (+12%)** | **0.220 ms** | **3250.0000 (+63%)** | **500.0000 (+25%)** | **0.0000** | **14689.45 KB (+65%)** |


## Jint.Benchmark.EvaluationBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|953.9 μs|7.83 μs|167.9688|0.0000|0.0000|690.8 KB|
| **New** |	|	| **1.153 ms (+24%)** | **0.0097 ms** | **179.6875 (+7%)** | **0.0000** | **0.0000** | **735.47 KB (+6%)** |


## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|1|939.5 ms|6.86 ms|42000.0000|0.0000|0.0000|168.97 MB|
| **New** |	|	| **1.438 s (+57%)** | **0.0053 s** | **79000.0000 (+88%)** | **0.0000** | **0.0000** | **317.7 MB (+88%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|269.0 ms|1.19 ms|44000.0000|1000.0000|0.0000|176.4 MB|
| **New** |	|	| **280.2 ms (+4%)** | **4.54 ms** | **48000.0000 (+9%)** | **13000.0000 (+1200%)** | **0.0000** | **192.83 MB (+9%)** |
| Old |Benchmark|500|203.2 ms|0.60 ms|38333.3333|7333.3333|0.0000|154.58 MB|
| **New** |	|	| **212.2 ms (+4%)** | **1.13 ms** | **42333.3333 (+10%)** | **3333.3333 (-55%)** | **0.0000** | **170.33 MB (+10%)** |